backend/requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
pydantic==2.5.3
facenet-pytorch==2.5.3
torch==2.1.2
torchvision==0.16.2
pillow==10.2.0
passlib==1.7.4
bcrypt==4.1.2
python-jose==3.3.0
python-dotenv
prometheus-client
sentry-sdk==1.39.1
fastapi-cache2==0.2.1
email-validator
python-multipart==0.0.6
sqlalchemy==2.0.25
pandas==2.1.4
redis==5.0.1
pydantic-settings==2.1.0
logging-utils
pytest==8.0.0
httpx==0.26.0
black
flake8
mypy
numpy==1.26.3
opencv-python==4.9.0.80
prometheus-fastapi-instrumentator==6.1.0
pytest-asyncio==0.23.4